[Fix] 장소 아이디값 수정
1. 기존에 파일,클래스명 이 placce이던것을 place로 수정함.
2. placeIdx:Long -> Int 작업 완료

+++ app/src/main/java/place/pic/ui/main/detail/liker/LikerUserListActivity.kt
@@ -12,14 +12,14 @@ import place.pic.ui.util.customEnqueue
 
 class LikerUserListActivity : AppCompatActivity() {
 
-    private var placeIdx:Long = 0L
+    private var placeIdx:Int = 0
 
     private lateinit var likerUserListAdapter:LikerUserListAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_liker_user_list)
-        placeIdx = intent.getLongExtra("placeIdx",0)
+        placeIdx = intent.getIntExtra("placeIdx",0)
         init()
     }
